refactor spectrogram generation functions to its own file
